made module folders for populating restaurants
